Add configuration file support.
Add default configuration which gets copied to local configuration if local
configuration does not exist (ex. first run).  Need to check whether creation
of local config directory is portable, using <sys/stat.h> macros and
g_mkdir_with_parents() function.

Add setting for whether the sidebar tabs should be forced to the bottom or left
at their default position.

Add setting to allow the Devhelp documentation tab to be located in the bottom
message window instead of the default main notebook location.

Need to write code to allow moving to message window and back to be done in
real-time rather than needing to reload the plugin or restart Geany.  For now,
a one-time message is shown indicating the need for a restart.

Create a new image, putting the Geany icon in place of the "D" on the cover
of the Devhelp book's icon.  Need to fix perspective of Geany icon to make it
look better.  Also, need to make this default home page look better all around.